requirements.txt
# Core Dependencies
streamlit==1.28.1
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1

# Optional: Enhanced Features (uncomment if using the enhanced app version)
# plotly==5.17.0

# Alternative lighter versions for deployment (if facing memory issues)
# tensorflow-cpu==2.13.0  # Use instead of tensorflow for CPU-only deployment
# opencv-python-headless==4.8.1.78  # Use instead of opencv-python for headless servers